pyproject.toml
[project]
name = "fast-api-beyond-crud"
version = "0.1.0"
description = "Book management API's for learning project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aioredis[hiredis]>=2.0.1",
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.128.0",
    "greenlet>=3.3.1",
    "pwdlib[argon2]>=0.3.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "redis>=7.1.0",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]
